Add entries about ECDSA leakage (#27)
* Add entries about ECDSA leakage

The ECDSA is fragile: even small leakages in nonce can help
recovering the private key. The leakage could be a result of a
faulty RNG or timing leak. The most unfortunate part is that this
fact is generally unknown. Therefore, I think this is a valuable
addition.

* Fix link with garykessler.net

Replace it with web archive.

Co-authored-by: Pavlo Farb <pavlo@cossacklabs.com>