Add classification of PAKE algorithms
PAKE algorithms have a long and unfortunate history. The task is
simple - exchange keys with a password or shared secret mixed in.
There are many requirements, and even more algorithms designed. But
most of them have their drawbacks, weak points, or are vulnerable
to particular attacks. This is still an area that requires research,
and this paper captures it.

So, I think it's a worthwhile addition as it raises awareness and
interest in such studies.