Fix links and update lychee in CI (#51)
* readme: Fix Cryptanalysis of block ciphers

The link was already on web.archive. However, we need to pin to
exact date because the latest snapshot is for 404 page.

* Fix Cache Attacks and Countermeasures

Here we have eprint link so let's use it. It's fast, free and has
HTTPS support :)

* Fix On the Security of Public Key Protocols

Unfortunately, the link is not working anymore and there are no free
resources with this work. So, let's use web.archive

* Use https for How to construct zero-knowledge proof

Maybe it will help CI because as for now it fails with this link.

* ci: Update lychee version

* Fix A Method for Obtaining Digital Signatures

The previous link was used unsafe TLS features, so lychee was not
happy about it.